Rename module to rule_registry
